+++ lib/screens/dialogs/battlefront_options_dialog.dart
@@ -21,7 +21,6 @@ class _BattlefrontOptionsDialogState extends State<BattlefrontOptionsDialog> {
   @override
   Widget build(BuildContext context) {
     return ContentDialog(
-      backgroundDismiss: false,
       constraints: const BoxConstraints(maxWidth: 700, maxHeight: 400),
       title: const Text("Battlefront Settings"),
       actions: [