+++ lib/screens/dialogs/walk_through/widgets/nexusmods_login.dart
@@ -68,7 +68,7 @@ class _NexusmodsLoginState extends State<NexusmodsLogin> {
         await box.put('nexusmods_login', true);
         if (!mounted) return;
         Navigator.of(context).pop();
-      } else if (!url.contains('nexusmods')) {
+      } else if (!url.contains('nexusmods') ||! url.startsWith("https://users.nexusmods.com") &&! showOverlay) {
         _controller.loadUrl('https://users.nexusmods.com/auth/sign_in');
       }
     });
@@ -86,7 +86,6 @@ class _NexusmodsLoginState extends State<NexusmodsLogin> {
   @override
   Widget build(BuildContext context) {
     return ContentDialog(
-      backgroundDismiss: false,
       constraints: BoxConstraints(maxWidth: browserOpen ? 1000 : 600, maxHeight: browserOpen ? 857 : 400),
       title: Text(translate('$prefix.title')),
       actions: [
@@ -138,7 +137,7 @@ class _NexusmodsLoginState extends State<NexusmodsLogin> {
                                     ),
                                     SizedBox(width: 15),
                                     Text(
-                                      'Loading...',
+                                      'Waiting for NexusMods...',
                                       style: TextStyle(fontSize: 19),
                                     ),
                                   ],