+++ lib/utils/services/frosty_service.dart
@@ -18,7 +18,7 @@ class FrostyService {
     String path = frostyPath ?? box.get('frostyPath');
     var r = await Process.run(
       '$path/FrostyModManager.exe',
-      launch ? ['-launch', profile ?? 'KyberModManager'] : [],
+      launch ? ['-launch', dynamicEnvEnabled ? (profile ?? 'KyberModManager') : 'KyberModManager'] : [],
       workingDirectory: path,
       includeParentEnvironment: true,
       runInShell: true,
@@ -71,7 +71,7 @@ class FrostyService {
       return false;
     }
 
-    return Version.parse(version.version.replaceAll('v', '')) < Version.parse('1.0.6-beta4');
+    return Version.parse(version.version.replaceAll('v', '')) < Version.parse(await ApiService.getLatestFrostyVersion());
   }
 
   static FrostyConfig getFrostyConfig([String? path]) {