+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 ## [1.0.8] - [-/-/-]
 
+- Faster loading of saved profile through symlinks.
+- Added the ability to manually select the BF2 path.
 - Added the faction selector on the host page.
 - Added an in-app browser for logging in.
+- Added button to manually select a Frosty pack when joining a server.
 - Fixed the infinite Cloudflare loading screen.
 - Fixed Frosty Collections for the cosmetics page.
+- Fixed that KMM couldn't connect to the API.
+- UI improvements.
 
 ## [1.0.7] - [13/06/22]